Patrick's comments:
Hi Jonathon,

Thanks for this - it's very useful. Clearly something is happening when no (smooth) filtering is applied which messes up results. Based on what you have shown, the last thing you tried (with smoothing) is what I would have expected intuitively.
It looks like  Sobel is not doing a very accurate angle estimation and the angles are "snapping" to angle values which are more or less vertical or horizontal (which is what the lines extracted from Canny look like). This is not what I would expect though - Sobel is meant to do some kind "aggregation" and the angles values in your CSV files seemed to occupy quite a wide range.
Looking at the Angle csv files between Approach 1 and Approach 3, the angles across the "band" (thick line) that sobel computes are *all consistent* in approach 3.  In the first (unfiltered) approach, the angles change a lot across the band, and the centre of the band  (probably what Canny returns) often seems to have an angle of 180 (or 0) This looks very much like an 'aliasing' issue....Because Canny extracts a 1 pixel 'curve' from the thick band given by Sobel, it only sees that centre angle (180/0 degrees). BY applying a Gaussian smoothing your ensure a that the angles across the entire band are all oriented consistently, as they should be intuitively, and so when Canny extracts the centre-most pixel in that band it reflects the true angle (or close too), and not some 'aliased'/quantized version (the 180/0 degrees). Well, that's my theory at least.

So, bottom line - use the third approach you outlined below. Yes, smoothing introduced some tiny (for narrow Gaussian e.g. 3x3) shift, but it's literally 1 pixel or  thereabouts (and canny does this anyway).
Sorry I didn't think of that aliasing issue, but I thought Sobel was 'smarter'. Of course, if Canny returned the angles, this would not be an issue! It may also be that Sobel has other parameters that control scale./aggregation - but I think if you have something that works now, stick with that. Generally everyone does a bit of smoothing :-) Of course, you can still refer to the original unsmoothed image (or do speckle denoise) if you need that in another stage,
Anyway, I think approach 3 gives what I would hope so that is great.

Patrick